third_party/WebKit/LayoutTests/fast/js/Promise-static-any.html

<!DOCTYPE html>
<html>
<head>
<script src="resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
description('Test Promise.');

window.jsTestIsAsync = true;
result = undefined;

var p1 = new Promise(function(r) { r.fulfill('p1'); });
var p2 = new Promise(function(r) { r.fulfill('p2'); });
var p3 = new Promise(function(r) { r.fulfill('p3'); });
var p4 = new Promise(function() {});
var p5 = new Promise(function() {});
var p6 = new Promise(function(r) { r.reject('p6'); });
var p7 = new Promise(function(r) { r.reject('p7'); });
var p8 = new Promise(function(r) { r.reject('p8'); });
var p9 = new Promise(function(r) { r.fulfill(p2); });

var case0 = Promise.any().then(function(result) {
  testPassed('Promise.any() is fulfilled.');
  window.result = result;
  shouldBe('result', 'undefined');
}, function() {
  testFailed('Promise.any() is rejected.');
});

var case1 = Promise.any(p4, p1, p6).then(function(result) {
  testPassed('Promise.any(p4, p1, p6) is fulfilled.');
  window.result = result;
  shouldBeEqualToString('result', 'p1');
}, function() {
  testFailed('Promise.any(p4, p1, p6) is rejected.');
});

var case2 = Promise.any(p4, p5).then(function(result) {
  testFailed('Promise.any(p4, p5) is fulfilled.');
}, function() {
  testFailed('Promise.any(p4, p5) is rejected.');
});

var case3 = Promise.any(p4, p6, p1).then(function(result) {
  testFailed('Promise.any(p4, p6, p1) is fulfilled.');
}, function(result) {
  testPassed('Promise.any(p4, p6, p1) is rejected.');
  window.result = result;
  shouldBeEqualToString('result', 'p6');
});

var case4 = Promise.any(p9).then(function(result) {
  testPassed('Promise.any(p9) is fulfilled.');
  window.result = result;
  shouldBeEqualToString('result', 'p2');
}, function() {
  testFailed('Promise.any(p9) is rejected.');
});

shouldBe('result', 'undefined');
Promise.every(case0, case1, case3, case4).then(finishJSTest, finishJSTest);

</script>
<script src="resources/js-test-post.js"></script>
</body>
</html>
