third_party/WebKit/LayoutTests/fast/js/Promise-static-any-expected.txt

Test Promise.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS result is undefined
PASS Promise.any() is fulfilled.
PASS result is undefined
PASS Promise.any(p4, p1, p6) is fulfilled.
PASS result is "p1"
PASS Promise.any(p4, p6, p1) is rejected.
PASS result is "p6"
PASS Promise.any(p9) is fulfilled.
PASS result is "p2"
PASS successfullyParsed is true

TEST COMPLETE
